Telemetry refactor
