+++ packages/types/src/codebase-index.ts
@@ -1,0 +1,37 @@
+import { z } from "zod"
+
+/**
+ * CodebaseIndexConfig
+ */
+
+export const codebaseIndexConfigSchema = z.object({
+	codebaseIndexEnabled: z.boolean().optional(),
+	codebaseIndexQdrantUrl: z.string().optional(),
+	codebaseIndexEmbedderProvider: z.enum(["openai", "ollama"]).optional(),
+	codebaseIndexEmbedderBaseUrl: z.string().optional(),
+	codebaseIndexEmbedderModelId: z.string().optional(),
+})
+
+export type CodebaseIndexConfig = z.infer<typeof codebaseIndexConfigSchema>
+
+/**
+ * CodebaseIndexModels
+ */
+
+export const codebaseIndexModelsSchema = z.object({
+	openai: z.record(z.string(), z.object({ dimension: z.number() })).optional(),
+	ollama: z.record(z.string(), z.object({ dimension: z.number() })).optional(),
+})
+
+export type CodebaseIndexModels = z.infer<typeof codebaseIndexModelsSchema>
+
+/**
+ * CdebaseIndexProvider
+ */
+
+export const codebaseIndexProviderSchema = z.object({
+	codeIndexOpenAiKey: z.string().optional(),
+	codeIndexQdrantApiKey: z.string().optional(),
+})
+
+export type CodebaseIndexProvider = z.infer<typeof codebaseIndexProviderSchema>
