+++ packages/types/src/experiment.ts
@@ -1,0 +1,26 @@
+import { z } from "zod"
+
+import type { Keys, Equals, AssertEqual } from "./type-fu.js"
+
+/**
+ * ExperimentId
+ */
+
+export const experimentIds = ["autoCondenseContext", "powerSteering"] as const
+
+export const experimentIdsSchema = z.enum(experimentIds)
+
+export type ExperimentId = z.infer<typeof experimentIdsSchema>
+
+/**
+ * Experiments
+ */
+
+export const experimentsSchema = z.object({
+	autoCondenseContext: z.boolean(),
+	powerSteering: z.boolean(),
+})
+
+export type Experiments = z.infer<typeof experimentsSchema>
+
+type _AssertExperiments = AssertEqual<Equals<ExperimentId, Keys<Experiments>>>
