+++ packages/types/src/index.ts
@@ -1,2 +1,15 @@
-export * from "./types.js"
 export * from "./api.js"
+export * from "./codebase-index.js"
+export * from "./experiment.js"
+export * from "./global-settings.js"
+export * from "./history.js"
+export * from "./ipc.js"
+export * from "./message.js"
+export * from "./mode.js"
+export * from "./model.js"
+export * from "./provider-settings.js"
+export * from "./telemetry.js"
+export * from "./terminal.js"
+export * from "./tool.js"
+export * from "./type-fu.js"
+export * from "./vscode.js"
