+++ src/services/telemetry/TelemetryService.ts
@@ -1,28 +1,35 @@
+import * as vscode from "vscode"
 import { ZodError } from "zod"
 
+import { TelemetryEventName } from "@roo-code/types"
+
 import { logger } from "../../utils/logging"
-import { PostHogClient, ClineProviderInterface } from "./PostHogClient"
+
+import { PostHogTelemetryClient } from "./clients/PostHogTelemetryClient"
+import { type TelemetryClient, type TelemetryPropertiesProvider } from "./types"
 
 /**
- * TelemetryService wrapper class that defers PostHogClient initialization
- * This ensures that we only create the PostHogClient after environment variables are loaded
+ * TelemetryService wrapper class that defers initialization.
+ * This ensures that we only create the various clients after environment
+ * variables are loaded.
  */
 class TelemetryService {
-	private client: PostHogClient | null = null
+	private clients: TelemetryClient[] = []
 	private initialized = false
 
 	/**
-	 * Initialize the telemetry service with the PostHog client
-	 * This should be called after environment variables are loaded
+	 * Initialize the telemetry client. This should be called after environment
+	 * variables are loaded.
 	 */
-	public initialize(): void {
+	public async initialize(context: vscode.ExtensionContext): Promise<void> {
 		if (this.initialized) {
 			return
 		}
 
+		this.initialized = true
+
 		try {
-			this.client = PostHogClient.getInstance()
-			this.initialized = true
+			this.clients.push(PostHogTelemetryClient.getInstance())
 		} catch (error) {
 			console.warn("Failed to initialize telemetry service:", error)
 		}
@@ -32,10 +39,10 @@ class TelemetryService {
 	 * Sets the ClineProvider reference to use for global properties
 	 * @param provider A ClineProvider instance to use
 	 */
-	public setProvider(provider: ClineProviderInterface): void {
-		// If client is initialized, pass the provider reference
+	public setProvider(provider: TelemetryPropertiesProvider): void {
+		// If client is initialized, pass the provider reference.
 		if (this.isReady) {
-			this.client!.setProvider(provider)
+			this.clients.forEach((client) => client.setProvider(provider))
 		}
 
 		logger.debug("TelemetryService: ClineProvider reference set")
@@ -47,7 +54,7 @@ class TelemetryService {
 	 * @returns Whether the service is ready to use
 	 */
 	private get isReady(): boolean {
-		return this.initialized && this.client !== null
+		return this.initialized && this.clients.length > 0
 	}
 
 	/**
@@ -59,19 +66,7 @@ class TelemetryService {
 			return
 		}
 
-		this.client!.updateTelemetryState(didUserOptIn)
-	}
-
-	/**
-	 * Captures a telemetry event if telemetry is enabled
-	 * @param event The event to capture with its properties
-	 */
-	public capture(event: { event: string; properties?: any }): void {
-		if (!this.isReady) {
-			return
-		}
-
-		this.client!.capture(event)
+		this.clients.forEach((client) => client.updateTelemetryState(didUserOptIn))
 	}
 
 	/**
@@ -79,45 +74,61 @@ class TelemetryService {
 	 * @param eventName The event name to capture
 	 * @param properties The event properties
 	 */
-	public captureEvent(eventName: string, properties?: any): void {
-		this.capture({ event: eventName, properties })
+	public captureEvent(eventName: TelemetryEventName, properties?: any): void {
+		if (!this.isReady) {
+			return
+		}
+
+		this.clients.forEach((client) => client.capture({ event: eventName, properties }))
 	}
 
-	// Task events convenience methods
 	public captureTaskCreated(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CREATED, { taskId })
+		this.captureEvent(TelemetryEventName.TASK_CREATED, { taskId })
 	}
 
 	public captureTaskRestarted(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.RESTARTED, { taskId })
+		this.captureEvent(TelemetryEventName.TASK_RESTARTED, { taskId })
 	}
 
 	public captureTaskCompleted(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.COMPLETED, { taskId })
+		this.captureEvent(TelemetryEventName.TASK_COMPLETED, { taskId })
 	}
 
 	public captureConversationMessage(taskId: string, source: "user" | "assistant"): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CONVERSATION_MESSAGE, { taskId, source })
+		this.captureEvent(TelemetryEventName.TASK_CONVERSATION_MESSAGE, { taskId, source })
+	}
+
+	public captureLlmCompletion(
+		taskId: string,
+		properties: {
+			inputTokens: number
+			outputTokens: number
+			cacheWriteTokens: number
+			cacheReadTokens: number
+			cost?: number
+		},
+	): void {
+		this.captureEvent(TelemetryEventName.LLM_COMPLETION, { taskId, ...properties })
 	}
 
 	public captureModeSwitch(taskId: string, newMode: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.MODE_SWITCH, { taskId, newMode })
+		this.captureEvent(TelemetryEventName.MODE_SWITCH, { taskId, newMode })
 	}
 
 	public captureToolUsage(taskId: string, tool: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.TOOL_USED, { taskId, tool })
+		this.captureEvent(TelemetryEventName.TOOL_USED, { taskId, tool })
 	}
 
 	public captureCheckpointCreated(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CHECKPOINT_CREATED, { taskId })
+		this.captureEvent(TelemetryEventName.CHECKPOINT_CREATED, { taskId })
 	}
 
 	public captureCheckpointDiffed(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CHECKPOINT_DIFFED, { taskId })
+		this.captureEvent(TelemetryEventName.CHECKPOINT_DIFFED, { taskId })
 	}
 
 	public captureCheckpointRestored(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CHECKPOINT_RESTORED, { taskId })
+		this.captureEvent(TelemetryEventName.CHECKPOINT_RESTORED, { taskId })
 	}
 
 	public captureContextCondensed(
@@ -126,7 +137,7 @@ class TelemetryService {
 		usedCustomPrompt?: boolean,
 		usedCustomApiHandler?: boolean,
 	): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CONTEXT_CONDENSED, {
+		this.captureEvent(TelemetryEventName.CONTEXT_CONDENSED, {
 			taskId,
 			isAutomaticTrigger,
 			...(usedCustomPrompt !== undefined && { usedCustomPrompt }),
@@ -135,32 +146,32 @@ class TelemetryService {
 	}
 
 	public captureSlidingWindowTruncation(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.SLIDING_WINDOW_TRUNCATION, { taskId })
+		this.captureEvent(TelemetryEventName.SLIDING_WINDOW_TRUNCATION, { taskId })
 	}
 
 	public captureCodeActionUsed(actionType: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.CODE_ACTION_USED, { actionType })
+		this.captureEvent(TelemetryEventName.CODE_ACTION_USED, { actionType })
 	}
 
 	public capturePromptEnhanced(taskId?: string): void {
-		this.captureEvent(PostHogClient.EVENTS.TASK.PROMPT_ENHANCED, { ...(taskId && { taskId }) })
+		this.captureEvent(TelemetryEventName.PROMPT_ENHANCED, { ...(taskId && { taskId }) })
 	}
 
 	public captureSchemaValidationError({ schemaName, error }: { schemaName: string; error: ZodError }): void {
 		// https://zod.dev/ERROR_HANDLING?id=formatting-errors
-		this.captureEvent(PostHogClient.EVENTS.ERRORS.SCHEMA_VALIDATION_ERROR, { schemaName, error: error.format() })
+		this.captureEvent(TelemetryEventName.SCHEMA_VALIDATION_ERROR, { schemaName, error: error.format() })
 	}
 
 	public captureDiffApplicationError(taskId: string, consecutiveMistakeCount: number): void {
-		this.captureEvent(PostHogClient.EVENTS.ERRORS.DIFF_APPLICATION_ERROR, { taskId, consecutiveMistakeCount })
+		this.captureEvent(TelemetryEventName.DIFF_APPLICATION_ERROR, { taskId, consecutiveMistakeCount })
 	}
 
 	public captureShellIntegrationError(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.ERRORS.SHELL_INTEGRATION_ERROR, { taskId })
+		this.captureEvent(TelemetryEventName.SHELL_INTEGRATION_ERROR, { taskId })
 	}
 
 	public captureConsecutiveMistakeError(taskId: string): void {
-		this.captureEvent(PostHogClient.EVENTS.ERRORS.CONSECUTIVE_MISTAKE_ERROR, { taskId })
+		this.captureEvent(TelemetryEventName.CONSECUTIVE_MISTAKE_ERROR, { taskId })
 	}
 
 	/**
@@ -168,7 +179,7 @@ class TelemetryService {
 	 * @param button The button that was clicked
 	 */
 	public captureTitleButtonClicked(button: string): void {
-		this.captureEvent("Title Button Clicked", { button })
+		this.captureEvent(TelemetryEventName.TITLE_BUTTON_CLICKED, { button })
 	}
 
 	/**
@@ -176,20 +187,16 @@ class TelemetryService {
 	 * @returns Whether telemetry is enabled
 	 */
 	public isTelemetryEnabled(): boolean {
-		return this.isReady && this.client!.isTelemetryEnabled()
+		return this.isReady && this.clients.some((client) => client.isTelemetryEnabled())
 	}
 
-	/**
-	 * Shuts down the PostHog client
-	 */
 	public async shutdown(): Promise<void> {
 		if (!this.isReady) {
 			return
 		}
 
-		await this.client!.shutdown()
+		this.clients.forEach((client) => client.shutdown())
 	}
 }
 
-// Export a singleton instance of the telemetry service wrapper
 export const telemetryService = new TelemetryService()
