+++ src/services/telemetry/index.ts
@@ -1,0 +1,2 @@
+export * from "./TelemetryService"
+export * from "./types"
